big renaming (from \NodaSoft\ReferencesOperation to \NodaSoft\Operation)
