+++ php/v2/src/Operation/InitialData/InitialData.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace NodaSoft\Operation\InitialData;
+
+interface InitialData
+{
+
+}
