+++ php/v2/src/Operation/Result/Result.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace NodaSoft\Operation\Result;
+
+interface Result
+{
+    public function toArray(): array;
+}
