--- php/v2/src/ReferencesOperation/Command/ReferencesOperationCommand.php
@@ -1,15 +1,0 @@
-<?php
-
-namespace NodaSoft\ReferencesOperation\Command;
-
-use NodaSoft\ReferencesOperation\InitialData\InitialData;
-use NodaSoft\ReferencesOperation\Result\ReferencesOperationResult;
-
-interface ReferencesOperationCommand
-{
-    public function execute(): ReferencesOperationResult;
-
-    public function setResult(ReferencesOperationResult $result): void;
-
-    public function setInitialData(InitialData $initialData): void;
-}
