--- php/v2/src/ReferencesOperation/InitialData/InitialData.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace NodaSoft\ReferencesOperation\InitialData;
-
-interface InitialData
-{
-
-}
