--- php/v2/src/ReferencesOperation/Result/ReferencesOperationResult.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace NodaSoft\ReferencesOperation\Result;
-
-interface ReferencesOperationResult
-{
-    public function toArray(): array;
-}
